backend/requirements.txt
fastapi==0.129.2
httpx==0.28.1
openai==2.21.0
postgrest==2.28.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
uvicorn==0.41.0
